show area progress from krs
