+++ vss-extension.json
@@ -2,7 +2,7 @@
     "manifestVersion": 1,
     "id": "okr-hub",
     "publisher": "nianwensong",
-    "version": "1.0.360",
+    "version": "1.0.361",
     "name": "OKR Hub",
     "description": "",
     "public": false,
